+++ protected/humhub/modules/content/widgets/richtext/extensions/link/LinkParserBlock.php
@@ -2,13 +2,14 @@
 
 namespace humhub\modules\content\widgets\richtext\extensions\link;
 
-use humhub\helpers\Html;
 use yii\base\Model;
 use yii\helpers\Url;
 
 /**
  * Link: <orig> = [<text>](<url> "<title>")
  * Image: <orig> = ![<text><alignment>](<url> "<title>" =<width>x<height>)
+ * Video: <orig> = ![<text><alignment>](<url> video controls autoplay muted loop =<width>x<height>)
+ * Audio: <orig> = ![<text><alignment>](<url> audio controls autoplay muted loop)
  */
 class LinkParserBlock extends Model
 {
@@ -21,6 +22,11 @@ class LinkParserBlock extends Model
     public const BLOCK_KEY_HEIGHT = 'height';
     public const BLOCK_KEY_CLASS = 'class';
     public const BLOCK_KEY_STYLE = 'style';
+    public const BLOCK_KEY_TYPE = 'type';
+    public const BLOCK_KEY_CONTROLS = 'controls';
+    public const BLOCK_KEY_AUTOPLAY = 'autoplay';
+    public const BLOCK_KEY_MUTED = 'muted';
+    public const BLOCK_KEY_LOOP = 'loop';
 
     /**
      * @var array
@@ -77,9 +83,41 @@ class LinkParserBlock extends Model
             }
         }
 
-        if ($this->hasOption(static::BLOCK_KEY_MD) && preg_match('/=(\d+)?x(\d+)?\)$/', (string) $this->block[static::BLOCK_KEY_MD], $size)) {
-            $this->setWidth($size[1] ?? null);
-            $this->setHeight($size[2] ?? null);
+        if ($this->hasOption(static::BLOCK_KEY_MD)) {
+            $origTag = (string)$this->block[static::BLOCK_KEY_MD];
+
+            if (preg_match('/\((.*?)\)/', $origTag, $options)) {
+                $options = preg_split('/\s+/', $options[1], -1, PREG_SPLIT_NO_EMPTY);
+                foreach ($options as $option) {
+                    switch ($option) {
+                        case 'video':
+                        case 'audio':
+                            $this->setType($option);
+                            break;
+                        case 'controls':
+                            $this->setControls(true);
+                            break;
+                        case 'autoplay':
+                            $this->setAutoplay(true);
+                            break;
+                        case 'muted':
+                            $this->setMuted(true);
+                            break;
+                        case 'loop':
+                            $this->setLoop(true);
+                            break;
+                    }
+                }
+            }
+
+            if (preg_match('/\([^)]+?\b(video|audio)\b[^)]*?\)/i', $origTag, $type)) {
+                $this->setType($type[1]);
+            }
+
+            if (preg_match('/=(\d+)?x(\d+)?\)$/', $origTag, $size)) {
+                $this->setWidth($size[1] ?? null);
+                $this->setHeight($size[2] ?? null);
+            }
         }
     }
 
@@ -193,6 +231,56 @@ class LinkParserBlock extends Model
         $this->block[static::BLOCK_KEY_STYLE] = array_merge($this->block[static::BLOCK_KEY_STYLE], $style);
     }
 
+    public function setType(string $type): void
+    {
+        $this->block[static::BLOCK_KEY_TYPE] = $type;
+    }
+
+    public function getType(): string
+    {
+        return $this->block[static::BLOCK_KEY_TYPE] ?? 'image';
+    }
+
+    public function setControls(bool $enabled): void
+    {
+        $this->block[static::BLOCK_KEY_CONTROLS] = $enabled;
+    }
+
+    public function getControls(): bool
+    {
+        return $this->block[static::BLOCK_KEY_CONTROLS] ?? false;
+    }
+
+    public function setAutoplay(bool $enabled): void
+    {
+        $this->block[static::BLOCK_KEY_AUTOPLAY] = $enabled;
+    }
+
+    public function getAutoplay(): bool
+    {
+        return $this->block[static::BLOCK_KEY_AUTOPLAY] ?? false;
+    }
+
+    public function setMuted(bool $enabled): void
+    {
+        $this->block[static::BLOCK_KEY_MUTED] = $enabled;
+    }
+
+    public function getMuted(): bool
+    {
+        return $this->block[static::BLOCK_KEY_MUTED] ?? false;
+    }
+
+    public function setLoop(bool $enabled): void
+    {
+        $this->block[static::BLOCK_KEY_LOOP] = $enabled;
+    }
+
+    public function getLoop(): bool
+    {
+        return $this->block[static::BLOCK_KEY_LOOP] ?? false;
+    }
+
     public function setBlock(string $text, string $url, ?string $title = null, $fileId = null)
     {
         $this->setUrl($url);
@@ -244,7 +332,7 @@ class LinkParserBlock extends Model
         return isset($this->block[$key]) && $this->block[$key] !== '' && $this->block[$key] !== [];
     }
 
-    public function renderImageAttributes(): string
+    public function getImageAttributes(): array
     {
         $attrs = [
             'src' => $this->getUrl(),
@@ -266,7 +354,43 @@ class LinkParserBlock extends Model
             $attrs['style'] = $this->getStyle();
         }
 
-        return Html::renderTagAttributes($attrs);
+        return $attrs;
+    }
+
+    public function getVideoAttributes(): array
+    {
+        $attrs = $this->getImageAttributes();
+        $attrs['title'] = $this->getText();
+        unset($attrs['alt']);
+
+        if ($this->hasOption(static::BLOCK_KEY_CONTROLS)) {
+            $attrs['controls'] = 'controls';
+        }
+        if ($this->hasOption(static::BLOCK_KEY_AUTOPLAY)) {
+            $attrs['autoplay'] = 'autoplay';
+        }
+        if ($this->hasOption(static::BLOCK_KEY_MUTED)) {
+            $attrs['muted'] = 'muted';
+        }
+        if ($this->hasOption(static::BLOCK_KEY_LOOP)) {
+            $attrs['loop'] = 'loop';
+        }
+
+        return $attrs;
+    }
+
+    public function getAudioAttributes(): array
+    {
+        $attrs = $this->getVideoAttributes();
+
+        if (isset($attrs['width'])) {
+            unset($attrs['width']);
+        }
+        if (isset($attrs['height'])) {
+            unset($attrs['height']);
+        }
+
+        return $attrs;
     }
 
 }