+++ protected/humhub/modules/like/notifications/views/mails/newLike.php
@@ -23,7 +23,7 @@ use humhub\widgets\mails\MailButtonList;
 $likedRecord = $viewable->getLikedRecord();
 ?>
 
-<?php $this->beginContent('@notification/views/layouts/mail.php', $_params_); ?>
+<?php $this->beginContent('@notification/views/layouts/mail.php') ?>
     <table width="100%" border="0" cellspacing="0" cellpadding="0" align="left">
         <tr>
             <td>