+++ CHANGELOG.md
@@ -1,8 +1,14 @@
 HumHub Changelog
 ================
 
-1.18.3 (TBD)
-------------
+1.18.4 (Unreleased)
+---------------------
+- Enh #8170: Handle controllers with using external modules
+- Enh #8176: Upgrade Twig package to v3.26.0
+- Enh #8150: Topic sidebar widget
+
+1.18.3 (May 18, 2026)
+---------------------
 - Fix #8079: Hide wall entry topics container when no topics
 - Fix #8075: Filtering by Global Topic in Dashboard
 - Fix #8081: Adds null check for removeItem and appendItem.
@@ -25,6 +31,7 @@ HumHub Changelog
 - Enh #8122: Update PHPspreadsheet
 - Enh #8112: Update Mobile-Detect library to 4.10.0
 - Fix #8125: Show oEmbed warning only for supported embed URLs
+- Enh #8131: Support video and audio tags by RichText converters
 - Fix #8134: Revert "Improve `<p>` style in RichText editor"
 - Fix #8133: Rebuild search index after file updating
 - Fix #8132: Improve RichText styles in email message
@@ -34,7 +41,10 @@ HumHub Changelog
 - Fix #8143: Fix updating of Space memberships for existing members
 - Fix #8145: Preserve login return URL for PJAX requests
 - Fix #8144: Restrict file view action
-- Enh #8150: Topic sidebar widget
+- Enh #8156: Enhance behavior for vertical videos and multiple video attachments
+- Fix #8148: Yii2 2.0.55 compatibility
+- Fix #8163: Fix remove all space members permission
+- Fix #8164: Fix space members list visibility
 
 1.18.2 (March 22, 2026)
 -----------------------